+++ sentry/src/main/java/io/sentry/ProfilingTraceData.java
@@ -74,7 +74,8 @@ public final class ProfilingTraceData implements JsonUnknown, JsonSerializable {
     this(new File("dummy"), NoOpTransaction.getInstance());
   }
 
-  public ProfilingTraceData(@NotNull File traceFile, @NotNull ITransaction transaction) {
+  public ProfilingTraceData(
+      final @NotNull File traceFile, final @NotNull ITransaction transaction) {
     this(
         traceFile,
         new ArrayList<>(),
@@ -98,24 +99,24 @@ public final class ProfilingTraceData implements JsonUnknown, JsonSerializable {
   }
 
   public ProfilingTraceData(
-      @NotNull File traceFile,
-      @NotNull List<ProfilingTransactionData> transactions,
-      @NotNull ITransaction transaction,
-      @NotNull String durationNanos,
-      int sdkInt,
-      @NotNull String cpuArchitecture,
-      @NotNull Callable<List<Integer>> deviceCpuFrequenciesReader,
-      @Nullable String deviceManufacturer,
-      @Nullable String deviceModel,
-      @Nullable String deviceOsVersion,
-      @Nullable Boolean deviceIsEmulator,
-      @Nullable String devicePhysicalMemoryBytes,
-      @Nullable String buildId,
-      @Nullable String versionName,
-      @Nullable String versionCode,
-      @Nullable String environment,
-      @NotNull String truncationReason,
-      @NotNull Map<String, ProfileMeasurement> measurementsMap) {
+      final @NotNull File traceFile,
+      final @NotNull List<ProfilingTransactionData> transactions,
+      final @NotNull ITransaction transaction,
+      final @NotNull String durationNanos,
+      final int sdkInt,
+      final @NotNull String cpuArchitecture,
+      final @NotNull Callable<List<Integer>> deviceCpuFrequenciesReader,
+      final @Nullable String deviceManufacturer,
+      final @Nullable String deviceModel,
+      final @Nullable String deviceOsVersion,
+      final @Nullable Boolean deviceIsEmulator,
+      final @Nullable String devicePhysicalMemoryBytes,
+      final @Nullable String buildId,
+      final @Nullable String versionName,
+      final @Nullable String versionCode,
+      final @Nullable String environment,
+      final @NotNull String truncationReason,
+      final @NotNull Map<String, ProfileMeasurement> measurementsMap) {
     this.traceFile = traceFile;
     this.cpuArchitecture = cpuArchitecture;
     this.deviceCpuFrequenciesReader = deviceCpuFrequenciesReader;
@@ -267,91 +268,91 @@ public final class ProfilingTraceData implements JsonUnknown, JsonSerializable {
     return measurementsMap;
   }
 
-  public void setAndroidApiLevel(int androidApiLevel) {
+  public void setAndroidApiLevel(final int androidApiLevel) {
     this.androidApiLevel = androidApiLevel;
   }
 
-  public void setCpuArchitecture(@NotNull String cpuArchitecture) {
+  public void setCpuArchitecture(final @NotNull String cpuArchitecture) {
     this.cpuArchitecture = cpuArchitecture;
   }
 
-  public void setDeviceLocale(@NotNull String deviceLocale) {
+  public void setDeviceLocale(final @NotNull String deviceLocale) {
     this.deviceLocale = deviceLocale;
   }
 
-  public void setDeviceManufacturer(@NotNull String deviceManufacturer) {
+  public void setDeviceManufacturer(final @NotNull String deviceManufacturer) {
     this.deviceManufacturer = deviceManufacturer;
   }
 
-  public void setDeviceModel(@NotNull String deviceModel) {
+  public void setDeviceModel(final @NotNull String deviceModel) {
     this.deviceModel = deviceModel;
   }
 
-  public void setDeviceOsBuildNumber(@NotNull String deviceOsBuildNumber) {
+  public void setDeviceOsBuildNumber(final @NotNull String deviceOsBuildNumber) {
     this.deviceOsBuildNumber = deviceOsBuildNumber;
   }
 
-  public void setDeviceOsVersion(@NotNull String deviceOsVersion) {
+  public void setDeviceOsVersion(final @NotNull String deviceOsVersion) {
     this.deviceOsVersion = deviceOsVersion;
   }
 
-  public void setDeviceIsEmulator(boolean deviceIsEmulator) {
+  public void setDeviceIsEmulator(final boolean deviceIsEmulator) {
     this.deviceIsEmulator = deviceIsEmulator;
   }
 
-  public void setDeviceCpuFrequencies(@NotNull List<Integer> deviceCpuFrequencies) {
+  public void setDeviceCpuFrequencies(final @NotNull List<Integer> deviceCpuFrequencies) {
     this.deviceCpuFrequencies = deviceCpuFrequencies;
   }
 
-  public void setDevicePhysicalMemoryBytes(@NotNull String devicePhysicalMemoryBytes) {
+  public void setDevicePhysicalMemoryBytes(final @NotNull String devicePhysicalMemoryBytes) {
     this.devicePhysicalMemoryBytes = devicePhysicalMemoryBytes;
   }
 
-  public void setTruncationReason(@NotNull String truncationReason) {
+  public void setTruncationReason(final @NotNull String truncationReason) {
     this.truncationReason = truncationReason;
   }
 
-  public void setTransactions(@NotNull List<ProfilingTransactionData> transactions) {
+  public void setTransactions(final @NotNull List<ProfilingTransactionData> transactions) {
     this.transactions = transactions;
   }
 
-  public void setBuildId(@NotNull String buildId) {
+  public void setBuildId(final @NotNull String buildId) {
     this.buildId = buildId;
   }
 
-  public void setTransactionName(@NotNull String transactionName) {
+  public void setTransactionName(final @NotNull String transactionName) {
     this.transactionName = transactionName;
   }
 
-  public void setDurationNs(@NotNull String durationNs) {
+  public void setDurationNs(final @NotNull String durationNs) {
     this.durationNs = durationNs;
   }
 
-  public void setVersionName(@NotNull String versionName) {
+  public void setVersionName(final @NotNull String versionName) {
     this.versionName = versionName;
   }
 
-  public void setVersionCode(@NotNull String versionCode) {
+  public void setVersionCode(final @NotNull String versionCode) {
     this.versionCode = versionCode;
   }
 
-  public void setTransactionId(@NotNull String transactionId) {
+  public void setTransactionId(final @NotNull String transactionId) {
     this.transactionId = transactionId;
   }
 
-  public void setTraceId(@NotNull String traceId) {
+  public void setTraceId(final @NotNull String traceId) {
     this.traceId = traceId;
   }
 
-  public void setProfileId(@NotNull String profileId) {
+  public void setProfileId(final @NotNull String profileId) {
     this.profileId = profileId;
   }
 
-  public void setEnvironment(@NotNull String environment) {
+  public void setEnvironment(final @NotNull String environment) {
     this.environment = environment;
   }
 
-  public void setSampledProfile(@Nullable String sampledProfile) {
+  public void setSampledProfile(final @Nullable String sampledProfile) {
     this.sampledProfile = sampledProfile;
   }
 
@@ -394,7 +395,7 @@ public final class ProfilingTraceData implements JsonUnknown, JsonSerializable {
   }
 
   @Override
-  public void serialize(@NotNull JsonObjectWriter writer, @NotNull ILogger logger)
+  public void serialize(final @NotNull JsonObjectWriter writer, final @NotNull ILogger logger)
       throws IOException {
     writer.beginObject();
     writer.name(JsonKeys.ANDROID_API_LEVEL).value(logger, androidApiLevel);
@@ -444,7 +445,7 @@ public final class ProfilingTraceData implements JsonUnknown, JsonSerializable {
   }
 
   @Override
-  public void setUnknown(@Nullable Map<String, Object> unknown) {
+  public void setUnknown(final @Nullable Map<String, Object> unknown) {
     this.unknown = unknown;
   }
 
@@ -453,7 +454,7 @@ public final class ProfilingTraceData implements JsonUnknown, JsonSerializable {
     @SuppressWarnings("unchecked")
     @Override
     public @NotNull ProfilingTraceData deserialize(
-        @NotNull JsonObjectReader reader, @NotNull ILogger logger) throws Exception {
+        final @NotNull JsonObjectReader reader, final @NotNull ILogger logger) throws Exception {
       reader.beginObject();
       ProfilingTraceData data = new ProfilingTraceData();
       Map<String, Object> unknown = null;